Catch the stack-extension signal on Windows Linux Sub-system (WLS)
The WLS seems to have a bug where it is raising the wrong signal code when the application attempt toread un umapped memorry segment. This is an issue since valgrind relies on this signal to detect that it needs to grow the stack of the application being instrumented.

This change introduces a naive hack to also catch the generic "from the kernel" code raised by WLS. Presumably this introduces a bug somewhere as someone much smarter that myself, who presumably understood the architecture, decided the check was necessary.

Running the regression suite on WLS now yields a much better result:

    == 225 tests, 9 stderr failures, 1 stdout failure, 0 stderrB failures, 0 stdoutB failures, 0 post failures ==
    memcheck/tests/addressable               (stderr)
    memcheck/tests/deep-backtrace            (stderr)
    memcheck/tests/file_locking              (stderr)
    memcheck/tests/leak_cpp_interior         (stderr)
    memcheck/tests/linux/getregset           (stdout)
    memcheck/tests/linux/getregset           (stderr)
    memcheck/tests/linux/stack_switch        (stderr)
    memcheck/tests/linux/timerfd-syscall     (stderr)
    memcheck/tests/pointer-trace             (stderr)
    memcheck/tests/sem                       (stderr